Use twoslash and shiki to create better experience for end user
Updates the websocketkit dependency to version 0.1.0-alpha.5 in package.json and pnpm-lock.yaml.

Moves and renames documentation files related to Pusher channels to a more general 'api' directory. The content of these files is also updated to better reflect usage and examples, including the addition of twoslash snippets and improved descriptions. The deprecated `api.mdx` file is removed.